ActiveX: List Plugin directy after installation of VLC in IE
The plugin is in Internet Explorer <= v9 not listed in "Tools" ->
"Manage Add-ons" after installation of VLC. Only when I run the plugin
on a page and confirm a security message, that appears, the plugin
appears permanently in the list. In IE10 (Win8) the plugin is listed
directly after the installation.

Close #8788